requirements.txt
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
requests>=2.31
tqdm>=4.66
python-dateutil>=2.9
pytz>=2024.1
yfinance>=0.2.40
pandas_market_calendars>=4.4

torch>=2.2
transformers>=4.44
datasets>=2.19
accelerate>=0.33
peft>=0.11
